Refactor: Organize project into feature-based directory structure
- Moved components into src/features/ for modularity.
- Updated all import paths to reflect new locations.
- Created src/components/ for generic UI components.
- Placed types in src/types/ and constants in src/constants.ts.
- Reorganized files as per professional React project standards.
