+++ src/features/trainer/BotTrainer.tsx
